⚙ Fix: 친구 초대 로직 챌린지 페이지 내로 변경
